debian/bin/gencontrol.py: Set encoding to UTF-8 globally
We open some, but not all, files with an explicit UTF-8 encoding.  One
of the open calls that I missed has just caused gencontrol.py to fail
instead a pbuilder environment.  Instead of continuing to set an
explicit encoding for each open call, use locale.setlocale to set it
globally.