Add a new script for generating debian changelog
This adds a new script for generating the changes of the new release.
The script takes one argument, that is the linux-firmware upstream
repository. This script retrieves the previous commit version of the
package by looking into the debian/changelog. Then it gets the HEAD
commit of the repository and compare both by using 'git log' in reverse
mode. Currently, this is a script that has been strongly inspired from
debian/bin/stable-update (linux.git). A plan to improve the script in
order to analyze touched files and prefix entries in the changelog with
the right packages prefix OR to refactorize the code with stable-update
from linux.git will be done later.