brcm80211: add GPD Pocket and Win UMPC config
At boot, the kernel looks for
brcm/brcmfmac4356-pcie.gpd-win-pocket.txt to perform runtime NVRAM
configuration of the Broadcom 4356 chip on GPD Pocket and
(presumably, though I don't have one to confirm) Win UMPCs. As of
late 2018, the file is present in upstream linux-firmware. While
GPL2+, it serves little purpose without the non-free firmware for
the same chip so include it in firmware-brcm80211 rather than
firmware-linux-free.