Add "minimal" firmware for AP+STA mode
Add a build of 7.45.241 with a reduced feature set that supports 19
clients in AP mode, while simultaneously allowing it to connect in STA
mode. This is intended for applications such as Internet-in-a-Box that
don't require advanced features.

It is up to the user or distribution to rename the file or create a
suitable symbolic link in order to use this cut-down firmware.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>