d/copyright: Fix paragraph order for files in go7007/ and radeon/
In case multiple Files paragraphs specify patterns matching the same
file, the license in the later paragraph applies.  This means that
when one paragraph applies to files that are exceptions to a more
general pattern in another paragraph, the general must appear before
the exceptional.

For each of go7007/ and radeon/, we had general and exceptional
paragraphs in the wrong order.  Swap them around.