ipw2x00, ivtv: Add environment variable to skip EULA on installation
In order to use the piuparts test in the common salsa-ci pipeline, we
need to be able to get past the EULA question for these packages.  We
cannot use debconf preseeding for it but we can set environment
variables.

We probably should not implement a single environment variable through
which users could pretend to agree to all future EULAs.  So instead
use the specific names
FIRMWARE_{IPW2X00,IVTV}_LICENSE_READ_AND_ACCEPTED and require a value
of 'yes'.