Add Pico W (CYW43439) firmware
Although the Pico W firmware is distributed by other means, this is the
canonical location for Raspberry Pi-specific firmwares. Add the current
CYW43439 firmware (with appended clm_blob), as copied from
https://github.com/georgerobotics/cyw43-driver/tree/main/firmware
Note that the license file (called "copyright") has been updated to
show that the Cypress section also applies to this firmware.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>