d/salsa-ci.yml: Set changelog suite to UNRELEASED to avoid lintian error
Lintian always reports at least a warning for the "+salsaci" suffix,
and for backports it reports an error.  Avoid this by setting the
suite to UNRELEASED, the same as we do in src:linux.

An alternative might be to put all such warnings/errors in
SALSA_CI_LINTIAN_SUPPRESS_TAGS.