d/rules.real: Run dh_movetousr if available
dh_movetousr will move files from /lib to /usr/lib in trixie and later
releases, and will be a no-op when it is added to bookworm-backports.
To support backports to older suites, check whether it is available
before running it.

Also update lintian-overrides to match files in either /lib/firmware
or /usr/lib/firmware.