qcom-soc: Re-fix lintian override for lib/firmware/qcom/apq8096/modem.mbn
The binary-with-bad-dynamic-table error is on ftp-master's autoreject
list, and ftp-master is running an ancient version of lintian due to
later performance regressions.  It is therefore critical that we keep
matching this error without brackets around the filename.  To keep CI
happy as well, we use '*' to match with or without brackets.

This was previously fixed by Cyril Brulebois in version 20221214-5 and
then broken by me in 20230515-4, but not noticed until I needed to do
a binary upload for 20230625-3~exp1.