ivtv: Remove obsolete Recommends: firmware-linux-nonfree
This recommendation was added in version 0.38 to ensure that
v4l-cx25840.fw would remain installed after it was moved to
firmware-linux-nonfree.

It's now recommending the wrong package (that file is in
firmware-misc-nonfree), and I don't think such recommendations need to
be kept beyond the next stable release.  So remove it now.