nvidia-graphics: Add versioned Breaks on initramfs-tools to avoid bug #1076539
initramfs-tools bug #1076539 (duplication of firmware files accessed
through directory symlinks) has now been fixed in unstable and in
bookworm (in point release 12.7).

This bug was specifically triggered by firmware-nvidia-graphics in
conjunction with plymouth or other package that adds graphics drivers
to the initramfs.

To ensure that a backport of firmware-nvidia-graphics doesn't trigger
this, add a versioned Breaks against unfixed versions of
initramfs-tools.