d/watch: Tighten tag regexp to work with upstream as a remote
When debian/watch uses the 'git' method, and the current directory is
a Git repository with the upstream URL as a remote, uscan will scan
the local repository rather than cloning from that URL.

Currently this doesn't work because we match tags with arbitary names
as versions, which would include Debian tagged versions.  Change the
regexp to only match upstream tags.