d/salsa-ci.yml: Use uscan option to speed up Git export with Files-Excluded
uscan currently does both:

1. Generate a tarball from Git, since the upstream for a Debian
   source package must be a tarball.
2. Generate a second tarball with file exclusions applied.

This is rather inefficient.  However, the latest version added the
option to make the first tarball uncompressed.  Since we don't store
the first tarball, this is fine for us.

Use the environment variable to enable this option, since (a) we don't
run uscan directly but through origtargz (b) this will be safely
ignored when building in an older suite where uscan doesn't support
this.