d/copyright: Group and sort exclusions for size reasons
We exclude many files to limit the size of the source and/or binary
packages, but there's no general comment to explain that.  Add that
comment above all those file patterns.

The file patterns are listed in roughly alphabetical order, with lots
of different comments about the reasons, but these reasons fall into a
small number of categories.  As has already been done with the
licensing reasons, group them by categories and sort the patterns
within each category.