Copy debian_linux package from src:linux
firmware-nonfree and firmware-free are now the only users of the
linux-support-* packages, and they are only used in generation of a
source package, not as a build-dependency.

We currently have to keep updating the package name and adjusting for
API changes there whenever they occur.  Since the package no longer
contains anything tied to the kernel version, the version/ABI suffix
could be removed from its name, but that would still leave the
occasional API changes.

After discussion in merge request kernel-team/linux!1427, it was
agreed to treat the debian_linux Python package as a copylib instead.
This will allow src:linux to remove the linux-support-* package and
src:firmware-nonfree to follow its API changes only when ready.

To support this:

- Add a update-python-lib script to update the debian_linxu package
  from src:linux.

- Use it to copy the debian_linux package from linux version
  6.12.20-1.

- d/README.source, d/copyright: Update for the above changes.