d/copyright: mediatek: Stop excluding '*_MT7961*_1a*.bin' files
Due to Salsa's previous artifact limitations, these files were excluded:
- mediatek/BT_RAM_CODE_MT7961_1a_2_hdr.bin    (483K)
- mediatek/WIFI_MT7961_patch_mcu_1a_2_hdr.bin ( 99K)
- mediatek/WIFI_RAM_CODE_MT7961_1a.bin        (784K)

But now that the artifact limits have been increased, we can stop
excluding those files.

The copyright for these files is the same as for other MT7961 and the
globs already cover them.
As they were not previously shipped, they do not need to be added to the
'usrmovemitigation' list.

Signed-off-by: Diederik de Haas <didi.debian@cknow.org>