Add brcmfmac module parameters for stability
It has been observed by third parties that in circumstances yet to be
determined, the onboard WiFi can crash. The principal cause of this
seems to be "internal roaming". Disabling this feature is apparently
an effective way to prevent these crashes.

In the same vein, disable some other firmware features - fwsup, sae and
dump_obss - to no obvious detriment.