d/b/gencontrol.py: Change template substitution to allow for Jinja2 templates
The Templates class supports Jinja2 templates, but only if we pass a
context into its get method.  Getting a template and then calling
self.substitute() doesn't work since this loses the information about
the template format.

I want to use Jinja2 for the metainfo.xml template, and in future we
might want to use it for other templates.  So convert all the uses of
self.templates.get() to pass the context instead of doing substitution
separately, and delete the substitute method.

Also change all hyphenated variable names to use underscores instead,
because Jinja2 doesn't allow hyphens in variable names.