ipw2x00: Remove bogus Homepage field
We currently set the Homepage for firmware-ipw2x00 to a list of 2
URLs, which was always questionable.  I will start adding homepage
URLs to the AppStream metadata, and we really can only put a single
URL there.

Additionally, neither of the URLs currently in this list works any
more.

So remove the field entirely for this binary package only.