Refactor firmware installation to use dh_install, avoiding E2BIG
When updating to 20250808, so many files will be selected for
inclusion in firmware-cirrus that the first for-loop in the
binary_generic expands to longer than 128 kiB and the execve() call to
will fail with E2BIG.

We can reduce the command lines by about 50% by including only the
canonical names in FILES and LINKS and finding the source or link
target again at installation time.  But firmware-cirrus or another
binary package may well grow to the point that we hit the limit again.

Solve this properly by replacing the FILES and LINKS variables with
separate files that dh_install can process:

- d/b/gencontrol.py:
  - Combine files_real and links into files_selected
  - Delete the unused links_rev dict
  - Do not set FILES or LINKS make variables
  - Generate firmware-(package).install files listing source and
    destination directory for each file and link
- d/rules: In maintainerclean rule, delete debian/*.install
- d/rules.real: Replace direct file installation with dh_install