Override lintian file-without-copyright-info warning for license texts
Most license texts do not include a license or copyright information
for themselves.  Despite the lack of that explicit meta-license, we
are clearly intended and (usually) required to copy them along with
the works they apply to.

The machine-readable copyright format currently doesn't specify how to
describe these files (bug #1111126).  Override the lintian warnings
for this.