d/copyright: Exclude some more unwanted files
We already excluded the binaries under atusb/, because their source
was elswehere.  That left a ChangeLog which we don't describe.
Simplify things by excluding that too.

We excluded atmel/wilc1000_*.bin because it's meant for use with
microcontrollers and other systems too small to run Debian..  Exclude
atmel/wilc3000_*.bin for the same reason.