d/rules: Only deduplicate files that are packaged
Bug #1118195 was caused by dedupe-firmware.sh symlinking a file that
we included in a binary package to an identical file that we didn't
yet include.  In general we would want to include both the identical
files in the same package, but we don't currently have a way to make
that happen automatically.

To avoid similar bugs in future, switch to running dedupe-firmware.sh
on each package installation directory, not on debian/build/install.

Also stop running dedupe-firmware.sh in the debian/control-real
recipe, and update the comment there.