d/rules: Check for broken symlinks among installed firmware
Now that we only deduplicate files within each binary package, that
should not create any broken symlinks.  But symlinks can also be
explicitly defined by WHENCE, so we could still include such a symlink
and not the file to points to.

Add a check for broken symlinks after running dh_install and
deduplicating.  For now, this excludes the intentional cross-package
symlinks from firmware-qcom-soc to firmware-atheros.