intel-misc: Include Lenovo ThinkPad X1 2-in-1 Gen 10 ISH firmware
Add the ISH firmware for a Lenovo ThinkPad model.  This is under
LENOVO/ish because it has a different licence, but there is a symlink
to it under intel/ish that we already include.

This broken symlink was caught by the check added in the previous
commit.

Also documented this file in debian/copyright.