debian/copyright: drop firmware for Qualcomm QDU100 (Closes: #1126794)
Upstream has dropped the kernel_boot.elf due to GPL licence violation.
The QDU100 chip is a special chip targeting base stations, it is
unlikely that Debian users will need firmware for that chip. Drop the
whole set of binaries for QDU100 from the firmwre-nonfree package,
reclaiming space on Qualcomm user systems.

The comment currently specifies both reasons, it can be removed with
20260110-1.

Link: https://gitlab.com/kernel-firmware/linux-firmware/-/commit/acf533d4b3d7a002113d66a7830dbb19abf23419
Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@oss.qualcomm.com>