d/: skip files picked up by hexagon-dsp-binaries
The hexagon-dsp-binaries source package is picking up Qualcomm DSP
firmware, drop those from the firmware-nonfree package. To facilitate
seamless upgrades, add an explicit dependency on that package. It is
important, because e.g. UCSI functionality on some of those machines
depends on the ADSP firmware. Later it can be reduced to Recommends.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@oss.qualcomm.com>